comment : common 이름 변경, BaseEntity 고려사항 작성
